# packages/api-contract/specs/swagger-app.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/upload/image": {
      "post": {
        "operationId": "UploadController_uploadImage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Upload an image",
        "tags": [
          "upload"
        ],
        "x-scope": [
          "app",
          "user"
        ]
      }
    },
    "/api/v1/upload/file": {
      "post": {
        "operationId": "UploadController_uploadFile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Upload a file",
        "tags": [
          "upload"
        ],
        "x-scope": [
          "app",
          "user"
        ]
      }
    },
    "/api/v1/upload/multiple": {
      "post": {
        "operationId": "UploadController_uploadMultiple",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Upload multiple files",
        "tags": [
          "upload"
        ],
        "x-scope": [
          "app",
          "user"
        ]
      }
    },
...
    "/api/v1/subscription/verify": {
      "post": {
        "operationId": "SubscriptionApiController_verify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPurchaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Purchase verified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyPurchaseResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify Google purchase and activate subscription",
        "tags": [
          "Subscription"
        ],
        "x-scope": [
          "app"
        ]
      }
    },
    "/api/v1/subscription/status": {
      "get": {
        "operationId": "SubscriptionApiController_getStatus",
        "parameters": [
          {
            "name": "sync",
            "required": false,
            "in": "query",
            "description": "true để buộc verify lại với Google Play thay vì đọc cache DB",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription status returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionStatusResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current subscription status",
        "tags": [
          "Subscription"
        ],
        "x-scope": [
          "app"
        ]
      }
    },
    "/api/v1/subscription/restore": {
      "post": {
        "operationId": "SubscriptionApiController_restore",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestorePurchaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription restored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestorePurchaseResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restore purchase and activate subscription",
        "tags": [
          "Subscription"
        ],
        "x-scope": [
          "app"
        ]
      }
    }
  },
  "info": {
    "title": "PP02 API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token (Bearer <your_token_here>)"
      }
    },
    "schemas": {
...
      "VerifyPurchaseDto": {
        "type": "object",
        "properties": {
          "platform": {
            "type": "string",
            "enum": [
              "ANDROID",
              "IOS"
            ],
            "example": "IOS"
          },
          "productId": {
            "type": "string",
            "example": "weekly_plan"
          },
          "purchaseToken": {
            "type": "string",
            "example": "purchase-token-string"
          }
        },
        "required": [
          "platform",
          "productId",
          "purchaseToken"
        ]
      },
      "VerifyPurchaseResponseDto": {
        "type": "object",
        "properties": {
          "isPremium": {
            "type": "boolean",
            "example": true
          },
          "plan": {
            "type": "string",
            "enum": [
              "WEEKLY",
              "MONTHLY",
              "YEARLY",
              "LIFETIME"
            ],
            "example": "WEEKLY"
          },
          "isTrial": {
            "type": "boolean",
            "example": true
          },
          "expiresAt": {
            "type": "object",
            "example": "2026-05-20T00:00:00.000Z",
            "nullable": true
          },
          "autoRenew": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "isPremium",
          "plan",
          "isTrial",
          "expiresAt",
          "autoRenew"
        ]
      },
      "SubscriptionStatusResponseDto": {
        "type": "object",
        "properties": {
          "isPremium": {
            "type": "boolean",
            "example": true
          },
          "plan": {
            "type": "string",
            "enum": [
              "WEEKLY",
              "MONTHLY",
              "YEARLY",
              "LIFETIME"
            ],
            "example": "YEARLY",
            "nullable": true
          },
          "isTrial": {
            "type": "boolean",
            "example": false
          },
          "expiresAt": {
            "type": "object",
            "example": "2026-12-01T00:00:00.000Z",
            "nullable": true
          },
          "autoRenew": {
            "type": "boolean",
            "example": true
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "EXPIRED",
              "CANCELED",
              "GRACE"
            ],
            "example": "ACTIVE",
            "nullable": true
          }
        },
        "required": [
          "isPremium",
          "plan",
          "isTrial",
          "expiresAt",
          "autoRenew",
          "status"
        ]
      },
      "RestorePurchaseDto": {
        "type": "object",
        "properties": {
          "purchaseToken": {
            "type": "string",
            "example": "purchase-token-string"
          },
          "productId": {
            "type": "string",
            "example": "yearly_plan"
          }
        },
        "required": [
          "purchaseToken",
          "productId"
        ]
      },
      "RestorePurchaseResponseDto": {
        "type": "object",
        "properties": {
          "isPremium": {
            "type": "boolean",
            "example": true
          },
          "expiresAt": {
            "type": "object",
            "example": "2026-12-01T00:00:00.000Z",
            "nullable": true
          }
        },
        "required": [
          "isPremium",
          "expiresAt"
        ]
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}